Rename nested-modals example to be use the filename format '<topic>-<subtopic>.rs'.
